// src/modules/auth/redux/auth-action-types.enum.ts
// Auth action types

export enum AuthActionTypes {
  LOGIN_REQUESTED = '@auth/LOGIN_REQUESTED',
  LOGIN_LOADING = '@auth/LOGIN_LOADING',
  LOGIN_FAILURE = '@auth/LOGIN_FAILURE',
  LOGIN_SUCCESS = '@auth/LOGIN_SUCCESS',

  REGISTER_REQUESTED = '@auth/REGISTER_REQUESTED',
  REGISTER_LOADING = '@auth/REGISTER_LOADING',
  REGISTER_FAILURE = '@auth/REGISTER_FAILURE',
  REGISTER_SUCCESS = '@auth/REGISTER_SUCCESS',

  LOGIN_SOCIAL_ACCOUNT_REQUESTED = '@auth/LOGIN_SOCIAL_ACCOUNT_REQUESTED',
  LOGIN_SOCIAL_ACCOUNT_LOADING = '@auth/LOGIN_SOCIAL_ACCOUNT_LOADING',
  LOGIN_SOCIAL_ACCOUNT_FAILURE = '@auth/LOGIN_SOCIAL_ACCOUNT_FAILURE',
  LOGIN_SOCIAL_ACCOUNT_SUCCESS = '@auth/LOGIN_SOCIAL_ACCOUNT_SUCCESS',

  REVOKE_TOKEN_REQUESTED = '@auth/REVOKE_TOKEN_REQUESTED',
  REVOKE_TOKEN_LOADING = '@auth/REVOKE_TOKEN_LOADING',
  REVOKE_TOKEN_FAILURE = '@auth/REVOKE_TOKEN_FAILURE',
  REVOKE_TOKEN_SUCCESS = '@auth/REVOKE_TOKEN_SUCCESS',

  SET_REGISTER_FORM_IS_DIRTY = '@auth/SET_REGISTER_FORM_IS_DIRTY',

  RESET_AUTH_STATE = 'RESET_AUTH_STATE',
}